refactor: move the files around to a more sensible web app structure
