+++ apps/server/pkg/models/README.md
@@ -3,3 +3,5 @@
 These models are treated as the database agnostic version of the data and what
 is returned publicly. Internally, each database driver may use these models or
 implement their own versions.
+
+These are available publicly so they can be relied upon by external applications.
